@@ -1,5 +1,5 @@
 module github.com/cespare/tabular
 
-go 1.17
+go 1.25
 
 require github.com/google/go-cmp v0.5.7
